main/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
psutil==5.9.5
gunicorn==22.0.0
pytest==8.3.3
psycopg[binary]==3.2.9
